perf(search): switch embedding model dtype from fp32 to uint8
Reduces model download from ~86MB to ~23MB with negligible quality
loss for cosine similarity search.

+++ src/resortSearch.ts
@@ -19,7 +19,7 @@ const initPromise = (async () => {
   try {
     const { pipeline } = await import('@huggingface/transformers')
     const extractor = await pipeline('feature-extraction', MODEL_ID, {
-      dtype: 'fp32',
+      dtype: 'uint8',
     })
     embedder = async (text: string): Promise<number[]> => {
       const output = await extractor(text, {
